requirements.txt
django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
python-dotenv
drf-spectacular